backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
websockets==12.0
python-dotenv==1.0.0
Pillow==10.1.0
aiofiles==23.2.1
pydantic-settings==2.0.3
